Remove all remaining Python code
