+++ av1an-cli/Cargo.toml
@@ -3,7 +3,15 @@ name = "av1an-cli"
 version = "0.1.0"
 edition = "2018"
 
+[[bin]]
+name = "av1an"
+path = "src/main.rs"
+
 [dependencies]
 clap = "3.0.0-beta.2"
 serde_json = "1.0.64"
 serde = { version = "1.0.126", features = ["serde_derive"] }
+av1an-core = { path="../av1an-core" }
+av1an-pyo3 = { path="../av1an-pyo3" }
+shlex = "1.0.0"
+ctrlc = "3.1.9"
