+++ av1an-pyo3/Cargo.toml
@@ -16,22 +16,16 @@ requires-dist = [
     "vapoursynth",
 ]
 
-[features]
-extension-module = ["pyo3/extension-module"]
-default = ["extension-module"]
-
 [lib]
 name = "av1an_pyo3"
-crate-type = ["cdylib"]
+# crate-type = ["cdylib"]
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-pyo3 = { version = "0.14.1" }
 once_cell = "1.8.0"
 chrono = "0.4.19"
 av1an-core = { path="../av1an-core" }
-av1an-cli = { path="../av1an-cli" }
 av1an-encoder-constructor = { path="../av1an-encoder-constructor" }
 av1an-scene-detection = { path="../av1an-scene-detection" }
 dict_derive = "0.3.1"
@@ -42,8 +36,8 @@ serde_json = "1.0.64"
 serde = { version = "1.0.126", features = ["serde_derive"] }
 crossbeam-channel = "0.5.1"
 crossbeam-utils = "0.8.5"
-ctrlc = "3.1.9"
 itertools = "0.10.1"
+anyhow = "1.0.42"
 
 [dependencies.tokio]
 version = "1"
