--- av1an-pyo3/pyproject.toml
@@ -1,27 +1,0 @@
-[build-system]
-build-backend = "maturin"
-requires = ["setuptools", "maturin>=0.11,<0.12"]
-
-[tool.maturin]
-# cargo-extra-args = ["--release"]
-bindings = "pyo3"
-manylinux = "off"
-
-#[tool.maturin.dependencies]
-#python = "^3.6"
-
-[package.metadata.maturin]
-requires-dist = [
-    "numpy",
-    "scenedetect[opencv]",
-    "opencv-python",
-    "scipy",
-    "vapoursynth",
-]
-
-classifiers = [
-    "Programming Language :: Rust",
-    "Programming Language :: Python",
-    "Programming Language :: Python :: 3",
-    "Operating System :: OS Independent",
-]
