+++ av1an-pyo3/src/lib.rs
@@ -1,6 +1,7 @@
 use av1an_core::vapoursynth;
 use av1an_core::{ChunkMethod, Encoder};
 
+use anyhow::anyhow;
 use once_cell::sync::Lazy;
 use regex::Regex;
 use serde::{Deserialize, Serialize};
@@ -26,39 +27,27 @@ use std::str::FromStr;
 use std::time::Instant;
 use std::{collections::hash_map::DefaultHasher, path::PathBuf};
 
-use dict_derive::FromPyObject;
-use pyo3::prelude::*;
-use pyo3::wrap_pyfunction;
-
-#[pyfunction]
 fn adapt_probing_rate(rate: usize, _frames: usize) -> usize {
   av1an_core::adapt_probing_rate(rate)
 }
 
-#[pyfunction]
-fn get_keyframes(source: &str) -> PyResult<Vec<usize>> {
+fn get_keyframes(source: &str) -> anyhow::Result<Vec<usize>> {
   let pt = Path::new(source);
   let kf = av1an_core::ffmpeg::get_keyframes(pt);
   Ok(kf)
 }
 
-#[pyfunction]
-fn hash_path(path: &str) -> String {
+pub fn hash_path(path: &str) -> String {
   let mut s = DefaultHasher::new();
   path.hash(&mut s);
   format!("{:x}", s.finish())[..7].to_string()
 }
 
-#[pyfunction]
 fn construct_target_quality_command(
-  encoder: &str,
+  encoder: Encoder,
   threads: &str,
   q: &str,
-) -> PyResult<Vec<String>> {
-  let encoder = av1an_encoder_constructor::Encoder::from_str(encoder).map_err(|_| {
-    pyo3::exceptions::PyTypeError::new_err(format!("Unknown or unsupported encoder '{}'", encoder))
-  })?;
-
+) -> anyhow::Result<Vec<String>> {
   Ok(
     encoder
       .construct_target_quality_command(threads.parse().unwrap(), q.to_string())
@@ -69,14 +58,15 @@ fn construct_target_quality_command(
 }
 
 /// Creates vs pipe file
-#[pyfunction]
-fn create_vs_file(temp: &str, source: &str, chunk_method: &str) -> PyResult<String> {
+
+fn create_vs_file(temp: &str, source: &str, chunk_method: &str) -> anyhow::Result<String> {
   // only for python code, remove if being called by rust
   let temp = Path::new(temp);
   let source = Path::new(source).canonicalize()?;
   let chunk_method = ChunkMethod::from_str(chunk_method)
     // TODO implement this in the FromStr implementation itself
-    .map_err(|_| pyo3::exceptions::PyTypeError::new_err("Invalid chunk method"))?;
+    .unwrap();
+
   let load_script_path = temp.join("split").join("loadscript.vpy");
 
   if load_script_path.exists() {
@@ -89,10 +79,7 @@ fn create_vs_file(temp: &str, source: &str, chunk_method: &str) -> PyResult<Stri
     match chunk_method {
       ChunkMethod::FFMS2 => "ffindex",
       ChunkMethod::LSMASH => "lwi",
-      _ =>
-        return Err(pyo3::exceptions::PyTypeError::new_err(
-          "Can only use vapoursynth chunk methods"
-        )),
+      _ => return Err(anyhow!("invalid chunk method")),
     }
   )));
 
@@ -125,48 +112,27 @@ core.{}({:?}, cachefile={:?}).set_output()",
   Ok(load_script_path.to_string_lossy().to_string())
 }
 
-#[pyfunction]
 fn get_ffmpeg_info() -> String {
   av1an_core::get_ffmpeg_info()
 }
 
-#[pyfunction]
 fn get_frame_types(file: String) -> Vec<String> {
   let input_file = Path::new(&file);
 
   av1an_core::ffmpeg::get_frame_types(input_file)
 }
 
-#[pyfunction]
-fn determine_workers(encoder: &str) -> PyResult<u64> {
-  Ok(av1an_core::determine_workers(
-    Encoder::from_str(encoder).map_err(|_| {
-      pyo3::exceptions::PyTypeError::new_err(format!(
-        "Unknown or unsupported encoder '{}'",
-        encoder
-      ))
-    })?,
-  ))
+fn determine_workers(encoder: Encoder) -> anyhow::Result<u64> {
+  Ok(av1an_core::determine_workers(encoder))
 }
 
-#[pyfunction]
-fn frame_probe_vspipe(source: &str, py: Python) -> PyResult<usize> {
-  let frames = py.allow_threads(|| av1an_core::vapoursynth::num_frames(Path::new(source)));
-  frames.map_err(|e| pyo3::exceptions::PyTypeError::new_err(format!("{}", e)))
-}
-
-#[pyfunction]
-fn frame_probe(source: &str, py: Python) -> PyResult<usize> {
-  if is_vapoursynth(source) {
-    frame_probe_vspipe(source, py)
-  } else {
-    // TODO evaluate vapoursynth script in-memory if ffms2 or lsmash exists
-    Ok(ffmpeg_get_frame_count(source))
-  }
+fn frame_probe_vspipe(source: &str) -> anyhow::Result<usize> {
+  let frames = av1an_core::vapoursynth::num_frames(Path::new(source));
+  frames
 }
 
 // same as frame_probe, but you can call it without the python GIL
-fn frame_probe_rust(source: &str) -> usize {
+fn frame_probe(source: &str) -> usize {
   if is_vapoursynth(source) {
     av1an_core::vapoursynth::num_frames(Path::new(source)).unwrap()
   } else {
@@ -175,30 +141,21 @@ fn frame_probe_rust(source: &str) -> usize {
   }
 }
 
-#[pyfunction]
 fn extract_audio(input: &str, temp: &str, audio_params: Vec<String>) {
   let input_path = Path::new(&input);
   let temp_path = Path::new(&temp);
   av1an_core::ffmpeg::extract_audio(input_path, temp_path, &audio_params);
 }
 
-#[pyfunction]
 fn ffmpeg_get_frame_count(source: &str) -> usize {
   av1an_core::ffmpeg::ffmpeg_get_frame_count(Path::new(source))
 }
 
-#[pyfunction]
-fn concatenate_ivf(input: &str, output: &str) -> PyResult<()> {
+fn concatenate_ivf(input: &str, output: &str) -> anyhow::Result<()> {
   av1an_core::concat::concat_ivf(Path::new(input), Path::new(output))
-    .map_err(|e| pyo3::exceptions::PyTypeError::new_err(format!("{}", e)))
 }
 
-#[pyfunction]
-fn concatenate_ffmpeg(temp: String, output: String, encoder: String) -> PyResult<()> {
-  let encoder = Encoder::from_str(&encoder).map_err(|_| {
-    pyo3::exceptions::PyTypeError::new_err(format!("Unknown or unsupported encoder '{}'", encoder))
-  })?;
-
+fn concatenate_ffmpeg(temp: String, output: String, encoder: Encoder) -> anyhow::Result<()> {
   let temp_path = Path::new(&temp);
   let output_path = Path::new(&output);
 
@@ -206,20 +163,17 @@ fn concatenate_ffmpeg(temp: String, output: String, encoder: String) -> PyResult
   Ok(())
 }
 
-#[pyfunction]
 fn extra_splits(split_locations: Vec<usize>, total_frames: usize, split_size: usize) -> Vec<usize> {
   av1an_core::split::extra_splits(split_locations, total_frames, split_size)
 }
 
-#[pyfunction]
-fn segment(input: &str, temp: &str, segments: Vec<usize>) -> PyResult<()> {
+fn segment(input: &str, temp: &str, segments: Vec<usize>) -> anyhow::Result<()> {
   let input = Path::new(&input);
   let temp = Path::new(&temp);
   av1an_core::split::segment(input, temp, &segments);
   Ok(())
 }
 
-#[pyfunction]
 fn process_inputs(input: Vec<String>) -> Vec<String> {
   let path_bufs: Vec<PathBuf> = input
     .into_iter()
@@ -236,73 +190,47 @@ fn process_inputs(input: Vec<String>) -> Vec<String> {
   out
 }
 
-#[pyfunction]
 fn write_scenes_to_file(
   scenes: Vec<usize>,
   frames: usize,
   scenes_path_string: &str,
-) -> PyResult<()> {
+) -> anyhow::Result<()> {
   let scene_path = PathBuf::from(scenes_path_string);
 
   av1an_core::split::write_scenes_to_file(&scenes, frames, &scene_path).unwrap();
   Ok(())
 }
 
-#[pyfunction]
 fn read_scenes_from_file(scenes_path_string: &str) -> (Vec<usize>, usize) {
   let scene_path = PathBuf::from(scenes_path_string);
 
   av1an_core::split::read_scenes_from_file(&scene_path).unwrap()
 }
 
-#[pyfunction]
-fn parse_args() -> String {
-  av1an_cli::parse_args()
-}
-
-#[pyfunction]
-fn default_args() -> String {
-  av1an_cli::default_args()
-}
-
-#[pyfunction]
 fn vmaf_auto_threads(workers: usize) -> usize {
   av1an_core::target_quality::vmaf_auto_threads(workers)
 }
 
-#[pyfunction]
-fn set_log(file: &str) -> PyResult<()> {
+fn set_log(file: &str) -> anyhow::Result<()> {
   av1an_core::logger::set_log(file).unwrap();
   Ok(())
 }
 
-#[pyfunction]
-fn log(msg: &str) -> PyResult<()> {
+fn log(msg: &str) -> anyhow::Result<()> {
   av1an_core::logger::log(msg);
   Ok(())
 }
 
-#[pyfunction]
-fn get_default_pass(encoder: &str) -> PyResult<usize> {
-  Ok(
-    av1an_encoder_constructor::Encoder::from_str(&encoder)
-      .unwrap()
-      .get_default_pass(),
-  )
+fn get_default_pass(encoder: &str) -> anyhow::Result<u8> {
+  Ok(Encoder::from_str(&encoder).unwrap().get_default_pass())
 }
 
-#[pyfunction]
-fn get_default_cq_range(encoder: &str) -> PyResult<(usize, usize)> {
-  Ok(
-    av1an_encoder_constructor::Encoder::from_str(&encoder)
-      .unwrap()
-      .get_default_cq_range(),
-  )
+fn get_default_cq_range(encoder: &str) -> anyhow::Result<(usize, usize)> {
+  Ok(Encoder::from_str(&encoder).unwrap().get_default_cq_range())
 }
 
-#[pyfunction]
-fn get_default_arguments(encoder: &str) -> PyResult<Vec<String>> {
-  let encoder = av1an_encoder_constructor::Encoder::from_str(&encoder).unwrap();
+fn get_default_arguments(encoder: &str) -> anyhow::Result<Vec<String>> {
+  let encoder = Encoder::from_str(&encoder).unwrap();
   let default_arguments = encoder.get_default_arguments();
 
   Ok(
@@ -313,9 +241,8 @@ fn get_default_arguments(encoder: &str) -> PyResult<Vec<String>> {
   )
 }
 
-#[pyfunction]
-fn help_command(encoder: &str) -> PyResult<Vec<String>> {
-  let encoder = av1an_encoder_constructor::Encoder::from_str(&encoder).unwrap();
+fn help_command(encoder: &str) -> anyhow::Result<Vec<String>> {
+  let encoder = Encoder::from_str(&encoder).unwrap();
   let help_command = encoder.help_command();
 
   Ok(
@@ -326,83 +253,81 @@ fn help_command(encoder: &str) -> PyResult<Vec<String>> {
   )
 }
 
-#[pyfunction]
-fn encoder_bin(encoder: &str) -> PyResult<String> {
-  Ok(
-    av1an_encoder_constructor::Encoder::from_str(&encoder)
-      .unwrap()
-      .encoder_bin()
-      .into(),
-  )
+fn encoder_bin(encoder: &str) -> anyhow::Result<String> {
+  Ok(Encoder::from_str(&encoder).unwrap().encoder_bin().into())
 }
 
-#[pyfunction]
-fn output_extension(encoder: &str) -> PyResult<String> {
+fn output_extension(encoder: &str) -> anyhow::Result<String> {
   Ok(
-    av1an_encoder_constructor::Encoder::from_str(&encoder)
+    Encoder::from_str(&encoder)
       .unwrap()
       .output_extension()
       .into(),
   )
 }
 
-#[pyfunction]
-fn compose_ffmpeg_pipe(params: Vec<String>) -> PyResult<Vec<String>> {
-  let res = av1an_encoder_constructor::compose_ffmpeg_pipe(params);
+fn compose_ffmpeg_pipe(params: Vec<String>) -> anyhow::Result<Vec<String>> {
+  let res = av1an_core::compose_ffmpeg_pipe(params);
   Ok(res)
 }
 
-#[pyfunction]
-fn compose_1_1_pass(encoder: String, params: Vec<String>, output: String) -> PyResult<Vec<String>> {
-  let enc = av1an_encoder_constructor::Encoder::from_str(&encoder).unwrap();
+fn compose_1_1_pass(
+  encoder: String,
+  params: Vec<String>,
+  output: String,
+) -> anyhow::Result<Vec<String>> {
+  let enc = Encoder::from_str(&encoder).unwrap();
   Ok(enc.compose_1_1_pass(params, output))
 }
 
-#[pyfunction]
-fn compose_1_2_pass(encoder: String, params: Vec<String>, fpf: String) -> PyResult<Vec<String>> {
-  let enc = av1an_encoder_constructor::Encoder::from_str(&encoder).unwrap();
+fn compose_1_2_pass(
+  encoder: String,
+  params: Vec<String>,
+  fpf: String,
+) -> anyhow::Result<Vec<String>> {
+  let enc = Encoder::from_str(&encoder).unwrap();
   Ok(enc.compose_1_2_pass(params, fpf))
 }
 
-#[pyfunction]
 fn compose_2_2_pass(
   encoder: String,
   params: Vec<String>,
   fpf: String,
   output: String,
-) -> PyResult<Vec<String>> {
-  let enc = av1an_encoder_constructor::Encoder::from_str(&encoder).unwrap();
+) -> anyhow::Result<Vec<String>> {
+  let enc = Encoder::from_str(&encoder).unwrap();
   Ok(enc.compose_2_2_pass(params, fpf, output))
 }
 
-#[pyfunction]
 fn find_aom_keyframes(fl: String, min_kf_length: usize) -> Vec<usize> {
   let file = PathBuf::from(fl);
   av1an_scene_detection::aom_kf::find_aom_keyframes(file, min_kf_length)
 }
 
-#[pyfunction]
 fn man_command(encoder: String, params: Vec<String>, q: usize) -> Vec<String> {
-  let enc = av1an_encoder_constructor::Encoder::from_str(&encoder).unwrap();
+  let enc = Encoder::from_str(&encoder).unwrap();
 
   enc.man_command(params, q)
 }
 
-#[pyfunction]
-fn match_line(encoder: &str, line: &str) -> PyResult<usize> {
-  let enc = av1an_encoder_constructor::Encoder::from_str(encoder).unwrap();
+fn match_line(encoder: &str, line: &str) -> anyhow::Result<usize> {
+  let enc = Encoder::from_str(encoder).unwrap();
 
   Ok(enc.match_line(line).unwrap())
 }
 
-#[pyfunction]
-fn weighted_search(num1: f64, vmaf1: f64, num2: f64, vmaf2: f64, target: f64) -> PyResult<usize> {
+fn weighted_search(
+  num1: f64,
+  vmaf1: f64,
+  num2: f64,
+  vmaf2: f64,
+  target: f64,
+) -> anyhow::Result<usize> {
   Ok(av1an_core::target_quality::weighted_search(
     num1, vmaf1, num2, vmaf2, target,
   ))
 }
 
-#[pyfunction]
 fn probe_cmd(
   encoder: String,
   temp: String,
@@ -413,8 +338,8 @@ fn probe_cmd(
   n_threads: String,
   video_params: Vec<String>,
   probe_slow: bool,
-) -> PyResult<(Vec<String>, Vec<String>)> {
-  let encoder = av1an_encoder_constructor::Encoder::from_str(&encoder).unwrap();
+) -> anyhow::Result<(Vec<String>, Vec<String>)> {
+  let encoder = Encoder::from_str(&encoder).unwrap();
   Ok(encoder.probe_cmd(
     temp,
     name,
@@ -427,61 +352,52 @@ fn probe_cmd(
   ))
 }
 
-#[pyfunction]
-pub fn get_percentile(scores: Vec<f64>, percent: f64) -> PyResult<f64> {
+pub fn get_percentile(scores: Vec<f64>, percent: f64) -> anyhow::Result<f64> {
   // pyo3 doesn't seem to support `mut` in function declarations, so this is necessary
   let mut scores = scores;
   Ok(av1an_core::get_percentile(&mut scores, percent))
 }
 
-#[pyfunction]
-pub fn read_weighted_vmaf(fl: String, percentile: f64) -> PyResult<f64> {
+pub fn read_weighted_vmaf(fl: String, percentile: f64) -> anyhow::Result<f64> {
   let file = PathBuf::from(fl);
   let val = av1an_core::read_weighted_vmaf(&file, percentile).unwrap();
   Ok(val)
 }
 
-#[pyfunction]
-pub fn init_progress_bar(len: u64) -> PyResult<()> {
+pub fn init_progress_bar(len: u64) -> anyhow::Result<()> {
   av1an_core::progress_bar::init_progress_bar(len).unwrap();
   Ok(())
 }
 
-#[pyfunction]
-pub fn update_bar(inc: u64) -> PyResult<()> {
+pub fn update_bar(inc: u64) -> anyhow::Result<()> {
   av1an_core::progress_bar::update_bar(inc).unwrap();
   Ok(())
 }
 
-#[pyfunction]
-pub fn finish_progress_bar() -> PyResult<()> {
+pub fn finish_progress_bar() -> anyhow::Result<()> {
   av1an_core::progress_bar::finish_progress_bar().unwrap();
   Ok(())
 }
 
-#[pyfunction]
 pub fn plot_vmaf_score_file(scores_file_string: String, plot_path_string: String) {
   let scores_file = PathBuf::from(scores_file_string);
   let plot_path = PathBuf::from(plot_path_string);
   av1an_core::vmaf::plot_vmaf_score_file(&scores_file, &plot_path).unwrap()
 }
 
-#[pyfunction]
-pub fn validate_vmaf(model: &str) -> PyResult<()> {
+pub fn validate_vmaf(model: &str) -> anyhow::Result<()> {
   av1an_core::vmaf::validate_vmaf(&model).unwrap();
   Ok(())
 }
 
-#[pyfunction]
-pub fn plot_vmaf(source: &str, output: &str) -> PyResult<()> {
+pub fn plot_vmaf(source: &str, output: &str) -> anyhow::Result<()> {
   let input = PathBuf::from(source);
   let out = PathBuf::from(output);
   av1an_core::vmaf::plot_vmaf(&input, &out).unwrap();
   Ok(())
 }
 
-#[pyfunction]
-pub fn interpolate_target_q(scores: Vec<(f64, u32)>, target: f64) -> PyResult<(f64, f64)> {
+pub fn interpolate_target_q(scores: Vec<(f64, u32)>, target: f64) -> anyhow::Result<(f64, f64)> {
   let q = av1an_core::target_quality::interpolate_target_q(scores.clone(), target).unwrap();
 
   let vmaf = av1an_core::target_quality::interpolate_target_vmaf(scores, q).unwrap();
@@ -489,12 +405,10 @@ pub fn interpolate_target_q(scores: Vec<(f64, u32)>, target: f64) -> PyResult<(f
   Ok((q, vmaf))
 }
 
-#[pyfunction]
-pub fn interpolate_target_vmaf(scores: Vec<(f64, u32)>, target: f64) -> PyResult<f64> {
+pub fn interpolate_target_vmaf(scores: Vec<(f64, u32)>, target: f64) -> anyhow::Result<f64> {
   Ok(av1an_core::target_quality::interpolate_target_vmaf(scores, target).unwrap())
 }
 
-#[pyfunction]
 pub fn log_probes(
   vmaf_cq_scores: Vec<(f64, u32)>,
   frames: u32,
@@ -503,7 +417,7 @@ pub fn log_probes(
   target_q: u32,
   target_vmaf: f64,
   skip: String,
-) -> PyResult<()> {
+) -> anyhow::Result<()> {
   av1an_core::target_quality::log_probes(
     vmaf_cq_scores,
     frames,
@@ -516,14 +430,13 @@ pub fn log_probes(
   Ok(())
 }
 
-#[pyfunction]
 pub fn av_scenechange_detect(
   input: &str,
   total_frames: usize,
   min_scene_len: usize,
   quiet: bool,
   is_vs: bool,
-) -> PyResult<Vec<usize>> {
+) -> anyhow::Result<Vec<usize>> {
   if !quiet {
     println!("Scene detection");
     av1an_core::progress_bar::init_progress_bar(total_frames as u64).unwrap();
@@ -540,29 +453,20 @@ pub fn av_scenechange_detect(
     },
     min_scene_len,
     is_vs,
-  )
-  .map_err(|e| {
-    pyo3::exceptions::PyChildProcessError::new_err(format!(
-      "Error in av-scenechange detection: {}",
-      e
-    ))
-  });
+  )?;
 
   let _ = av1an_core::progress_bar::finish_progress_bar();
 
-  if let Ok(ref mut frames) = frames {
-    if frames[0] == 0 {
-      // TODO refactor the chunk creation to not require this
-      // Currently, this is required for compatibility with create_video_queue_vs
-      frames.remove(0);
-    }
+  if frames[0] == 0 {
+    // TODO refactor the chunk creation to not require this
+    // Currently, this is required for compatibility with create_video_queue_vs
+    frames.remove(0);
   }
 
-  frames
+  Ok(frames)
 }
 
-#[pyfunction]
-fn is_vapoursynth(s: &str) -> bool {
+pub fn is_vapoursynth(s: &str) -> bool {
   [".vpy", ".py"].iter().any(|ext| s.ends_with(ext))
 }
 
@@ -679,7 +583,7 @@ impl<'a> Queue<'a> {
   }
 
   fn frame_check_output(&self, chunk: &Chunk, expected_frames: usize) -> usize {
-    let actual_frames = frame_probe_rust(&chunk.output_path());
+    let actual_frames = frame_probe(&chunk.output_path());
 
     if actual_frames != expected_frames {
       let msg = format!(
@@ -694,7 +598,6 @@ impl<'a> Queue<'a> {
   }
 }
 
-#[pyclass]
 struct TargetQuality {
   vmaf_res: String,
   vmaf_filter: String,
@@ -713,9 +616,7 @@ struct TargetQuality {
   probe_slow: bool,
 }
 
-#[pymethods]
 impl TargetQuality {
-  #[new]
   fn new(project: &Project) -> Self {
     Self {
       vmaf_res: project
@@ -981,28 +882,18 @@ Chunk: {}
   Ok(())
 }
 
-#[pyclass(dict)]
 #[derive(Default, Debug, Serialize, Deserialize, Clone)]
 struct Chunk {
-  #[pyo3(get, set)]
   temp: String,
-  #[pyo3(get, set)]
   index: usize,
-  #[pyo3(get, set)]
   ffmpeg_gen_cmd: Vec<String>,
-  #[pyo3(get, set)]
   output_ext: String,
-  #[pyo3(get, set)]
   size: usize,
-  #[pyo3(get, set)]
   frames: usize,
-  #[pyo3(get, set)]
   per_shot_target_quality_cq: Option<u32>,
 }
 
-#[pymethods]
 impl Chunk {
-  #[new]
   fn new(
     temp: String,
     index: usize,
@@ -1022,17 +913,14 @@ impl Chunk {
     }
   }
 
-  #[getter]
   fn name(&self) -> String {
     format!("{:05}", self.index)
   }
 
-  #[getter]
   fn output(&self) -> String {
     self.output_path()
   }
 
-  #[getter]
   fn output_path(&self) -> String {
     Path::new(&self.temp)
       .join("encode")
@@ -1043,7 +931,6 @@ impl Chunk {
   }
 }
 
-#[pyfunction]
 fn save_chunk_queue(temp: &str, chunk_queue: Vec<Chunk>) {
   let mut file = File::create(Path::new(temp).join("chunks.json")).unwrap();
 
@@ -1052,99 +939,56 @@ fn save_chunk_queue(temp: &str, chunk_queue: Vec<Chunk>) {
     .unwrap();
 }
 
-#[pyclass(dict)]
-#[derive(Default, FromPyObject)]
-struct Project {
-  #[pyo3(get, set)]
-  frames: usize,
-  #[pyo3(get, set)]
-  is_vs: bool,
+#[derive(Default)]
+pub struct Project {
+  pub frames: usize,
+  pub is_vs: bool,
 
-  #[pyo3(get, set)]
-  input: String,
-  #[pyo3(get, set)]
-  temp: String,
-  #[pyo3(get, set)]
-  output_file: String,
-  #[pyo3(get, set)]
-  mkvmerge: bool,
-  #[pyo3(get, set)]
-  output_ivf: bool,
-  #[pyo3(get, set)]
-  webm: bool,
+  pub input: String,
+  pub temp: String,
+  pub output_file: String,
+  pub mkvmerge: bool,
+  pub output_ivf: bool,
+  pub webm: bool,
 
-  #[pyo3(get, set)]
-  chunk_method: Option<String>,
-  #[pyo3(get, set)]
-  scenes: Option<String>,
-  #[pyo3(get, set)]
-  split_method: String,
-  #[pyo3(get, set)]
-  extra_split: usize,
-  #[pyo3(get, set)]
-  min_scene_len: usize,
+  pub chunk_method: Option<String>,
+  pub scenes: Option<String>,
+  pub split_method: String,
+  pub extra_splits_len: Option<usize>,
+  pub min_scene_len: usize,
 
-  #[pyo3(get, set)]
-  passes: u8,
-  #[pyo3(get, set)]
-  video_params: Vec<String>,
-  #[pyo3(get, set)]
-  encoder: String,
-  #[pyo3(get, set)]
-  workers: usize,
+  pub passes: u8,
+  pub video_params: Vec<String>,
+  pub encoder: String,
+  pub workers: usize,
 
   // FFmpeg params
-  #[pyo3(get, set)]
-  ffmpeg_pipe: Vec<String>,
-  #[pyo3(get, set)]
-  ffmpeg: Vec<String>,
-  #[pyo3(get, set)]
-  audio_params: Vec<String>,
-  #[pyo3(get, set)]
-  pix_format: String,
+  pub ffmpeg_pipe: Vec<String>,
+  pub ffmpeg: Vec<String>,
+  pub audio_params: Vec<String>,
+  pub pix_format: String,
 
-  #[pyo3(get, set)]
-  quiet: bool,
-  #[pyo3(get, set)]
-  logging: String,
-  #[pyo3(get, set)]
-  resume: bool,
-  #[pyo3(get, set)]
-  keep: bool,
-  #[pyo3(get, set)]
-  force: bool,
+  pub quiet: bool,
+  pub logging: String,
+  pub resume: bool,
+  pub keep: bool,
 
-  #[pyo3(get, set)]
-  vmaf: bool,
-  #[pyo3(get, set)]
-  vmaf_path: Option<String>,
-  #[pyo3(get, set)]
-  vmaf_res: Option<String>,
+  pub vmaf: bool,
+  pub vmaf_path: Option<String>,
+  pub vmaf_res: Option<String>,
 
-  #[pyo3(get, set)]
-  concat: String,
+  pub concat: String,
 
-  #[pyo3(get, set)]
-  target_quality: Option<f32>,
-  #[pyo3(get, set)]
-  target_quality_method: Option<String>,
-  #[pyo3(get, set)]
-  probes: u32,
-  #[pyo3(get, set)]
-  probe_slow: bool,
-  #[pyo3(get, set)]
-  min_q: Option<u32>,
-  #[pyo3(get, set)]
-  max_q: Option<u32>,
-  // TODO: Refactor into just bool. Option<bool> is only used for Python interop.
-  #[pyo3(get, set)]
-  vmaf_plots: Option<bool>,
-  #[pyo3(get, set)]
-  probing_rate: u32,
-  #[pyo3(get, set)]
-  n_threads: Option<u32>,
-  #[pyo3(get, set)]
-  vmaf_filter: Option<String>,
+  pub target_quality: Option<f32>,
+  pub target_quality_method: Option<String>,
+  pub probes: u32,
+  pub probe_slow: bool,
+  pub min_q: Option<u32>,
+  pub max_q: Option<u32>,
+
+  pub probing_rate: u32,
+  pub n_threads: Option<u32>,
+  pub vmaf_filter: Option<String>,
 }
 
 static HELP_REGEX: Lazy<Regex> = Lazy::new(|| Regex::new(r"\s+(-\w+|(?:--\w+(?:-\w+)*))").unwrap());
@@ -1172,7 +1016,6 @@ fn suggest_fix(wrong_arg: &str, arg_dictionary: &HashSet<String>) -> Option<Stri
     .map(|(s, _)| (*s).to_owned())
 }
 
-#[pyfunction]
 fn read_chunk_queue(temp: &str) -> Vec<Chunk> {
   let contents = fs::read_to_string(Path::new(temp).join("chunks.json")).unwrap();
 
@@ -1321,7 +1164,7 @@ Chunk: {}
 {}",
           returncode,
           c.index,
-          encoder_history.iter().join(" ")
+          encoder_history.iter().join("\n")
         ));
       }
     }
@@ -1330,22 +1173,20 @@ Chunk: {}
   }
 }
 
-#[pymethods]
 impl Project {
-  #[new]
   fn new(project: Project) -> Self {
     project
   }
 
-  fn get_frames(&mut self, py: Python) -> usize {
+  fn get_frames(&mut self) -> usize {
     if self.frames != 0 {
       return self.frames;
     }
 
     self.frames = if self.is_vs {
-      py.allow_threads(|| vapoursynth::num_frames(Path::new(&self.input)).unwrap())
+      vapoursynth::num_frames(Path::new(&self.input)).unwrap()
     } else {
-      if ["vs_ffms2", "vs_lsmash"].contains(&self.chunk_method.as_ref().unwrap().as_str()) {
+      if ["ffms2", "lsmash"].contains(&self.chunk_method.as_ref().unwrap().as_str()) {
         let vs = if self.is_vs {
           self.input.clone()
         } else {
@@ -1356,7 +1197,7 @@ impl Project {
           )
           .unwrap()
         };
-        let fr = py.allow_threads(|| vapoursynth::num_frames(Path::new(&vs)).unwrap());
+        let fr = vapoursynth::num_frames(Path::new(&vs)).unwrap();
         if fr > 0 {
           fr
         } else {
@@ -1370,12 +1211,10 @@ impl Project {
     self.frames
   }
 
-  fn select_best_chunking_method(&mut self, py: Python) {
-    // You have to wrap vapoursynth calls with `allow_threads`, otherwise
-    // a fatal Python interpreter error occurs relating to the GIL state.
-    let chunk_method = py.allow_threads(|| av1an_core::vapoursynth::select_chunk_method().unwrap());
+  fn select_best_chunking_method(&mut self) {
+    let chunk_method = av1an_core::vapoursynth::select_chunk_method().unwrap();
 
-    self.chunk_method = Some(chunk_method.to_string());
+    self.chunk_method = Some(<&'static str>::from(chunk_method).to_string());
   }
 
   /// returns a list of valid parameters
@@ -1437,13 +1276,13 @@ impl Project {
     }
   }
 
-  fn startup_check(&mut self, py: Python) -> PyResult<()> {
+  pub fn startup_check(&mut self) -> anyhow::Result<()> {
     if ["rav1e", "aom", "svt_av1", "vpx"].contains(&self.encoder.as_str()) && self.output_ivf {
       panic!(".ivf only supports VP8, VP9, and AV1");
     }
 
     if self.chunk_method.is_none() {
-      self.select_best_chunking_method(py);
+      self.select_best_chunking_method();
     }
 
     assert!(
@@ -1510,12 +1349,12 @@ impl Project {
     Ok(())
   }
 
-  fn create_encoding_queue(&mut self, splits: Vec<usize>, py: Python) -> Vec<Chunk> {
+  fn create_encoding_queue(&mut self, splits: Vec<usize>) -> Vec<Chunk> {
     let mut chunks = match self.chunk_method.as_ref().unwrap().as_str() {
-      "vs_ffms2" | "vs_lsmash" => self.create_video_queue_vs(splits, py),
-      "hybrid" => self.create_video_queue_hybrid(splits, py),
-      "select" => self.create_video_queue_select(splits, py),
-      "segment" => self.create_video_queue_segment(splits, py),
+      "ffms2" | "lsmash" => self.create_video_queue_vs(splits),
+      "hybrid" => self.create_video_queue_hybrid(splits),
+      "select" => self.create_video_queue_select(splits),
+      "segment" => self.create_video_queue_segment(splits),
       _ => unreachable!(),
     };
 
@@ -1541,9 +1380,9 @@ impl Project {
 
   // If we are not resuming, then do scene detection. Otherwise: get scenes from
   // scenes.json and return that.
-  fn split_routine(&mut self, py: Python) -> Vec<usize> {
+  fn split_routine(&mut self) -> Vec<usize> {
     // TODO make self.frames impossible to misuse
-    let _ = self.get_frames(py);
+    let _ = self.get_frames();
 
     let scene_file = Path::new(&self.temp).join("scenes.json");
 
@@ -1656,8 +1495,8 @@ impl Project {
     }
   }
 
-  fn create_video_queue_vs(&mut self, splits: Vec<usize>, py: Python) -> Vec<Chunk> {
-    let last_frame = self.get_frames(py);
+  fn create_video_queue_vs(&mut self, splits: Vec<usize>) -> Vec<Chunk> {
+    let last_frame = self.get_frames();
 
     let mut split_locs = vec![0];
     split_locs.extend(splits);
@@ -1691,8 +1530,8 @@ impl Project {
     chunk_queue
   }
 
-  fn create_video_queue_select(&mut self, splits: Vec<usize>, py: Python) -> Vec<Chunk> {
-    let last_frame = self.get_frames(py);
+  fn create_video_queue_select(&mut self, splits: Vec<usize>) -> Vec<Chunk> {
+    let last_frame = self.get_frames();
 
     let mut split_locs = vec![0];
     split_locs.extend(splits);
@@ -1715,7 +1554,7 @@ impl Project {
     chunk_queue
   }
 
-  fn create_video_queue_segment(&mut self, splits: Vec<usize>, py: Python) -> Vec<Chunk> {
+  fn create_video_queue_segment(&mut self, splits: Vec<usize>) -> Vec<Chunk> {
     let _ = log("Split video");
     segment(&self.input, &self.temp, splits).unwrap();
     let _ = log("Split done");
@@ -1731,20 +1570,18 @@ impl Project {
     let chunk_queue: Vec<Chunk> = queue_files
       .iter()
       .enumerate()
-      .map(|(index, file)| {
-        self.create_chunk_from_segment(index, file.as_path().to_str().unwrap(), py)
-      })
+      .map(|(index, file)| self.create_chunk_from_segment(index, file.as_path().to_str().unwrap()))
       .collect();
 
     chunk_queue
   }
 
-  fn create_video_queue_hybrid(&mut self, split_locations: Vec<usize>, py: Python) -> Vec<Chunk> {
+  fn create_video_queue_hybrid(&mut self, split_locations: Vec<usize>) -> Vec<Chunk> {
     let keyframes = get_keyframes(&self.input).unwrap();
 
     let mut splits = vec![0];
     splits.extend(split_locations);
-    splits.push(self.get_frames(py));
+    splits.push(self.get_frames());
 
     let segments_set: HashSet<(usize, usize)> = splits
       .iter()
@@ -1796,7 +1633,7 @@ impl Project {
     chunk_queue
   }
 
-  fn create_chunk_from_segment(&mut self, index: usize, file: &str, py: Python) -> Chunk {
+  fn create_chunk_from_segment(&mut self, index: usize, file: &str) -> Chunk {
     let ffmpeg_gen_cmd = vec![
       "ffmpeg".into(),
       "-y".into(),
@@ -1819,7 +1656,7 @@ impl Project {
 
     Chunk {
       temp: self.temp.clone(),
-      frames: self.get_frames(py),
+      frames: self.get_frames(),
       ffmpeg_gen_cmd,
       output_ext,
       index,
@@ -1828,7 +1665,7 @@ impl Project {
     }
   }
 
-  fn load_or_gen_chunk_queue(&mut self, splits: Vec<usize>, py: Python) -> Vec<Chunk> {
+  fn load_or_gen_chunk_queue(&mut self, splits: Vec<usize>) -> Vec<Chunk> {
     if self.resume {
       let mut chunks = read_chunk_queue(&self.temp);
 
@@ -1842,37 +1679,31 @@ impl Project {
 
       chunks
     } else {
-      let chunks = self.create_encoding_queue(splits, py);
+      let chunks = self.create_encoding_queue(splits);
       save_chunk_queue(&self.temp, chunks.clone());
       chunks
     }
   }
 
-  fn encode_file(mut _self: PyRefMut<Self>, py: Python) {
-    ctrlc::set_handler(|| {
-      println!("Stopped");
-      std::process::exit(0);
-    })
-    .unwrap();
+  pub fn encode_file(&mut self) {
+    let _ = log(format!("File hash: {}", hash_path(&self.input)).as_str());
 
-    let _ = log(format!("File hash: {}", hash_path(&_self.input)).as_str());
+    let done_path = Path::new(&self.temp).join("done.json");
 
-    let done_path = Path::new(&_self.temp).join("done.json");
+    self.resume = self.resume && done_path.exists();
 
-    _self.resume = _self.resume && done_path.exists();
-
-    if !_self.resume && Path::new(&_self.temp).is_dir() {
-      fs::remove_dir_all(&_self.temp).unwrap();
+    if !self.resume && Path::new(&self.temp).is_dir() {
+      fs::remove_dir_all(&self.temp).unwrap();
     }
 
-    let _ = match fs::create_dir_all(Path::new(&_self.temp).join("split")) {
+    let _ = match fs::create_dir_all(Path::new(&self.temp).join("split")) {
       Ok(_) => {}
       Err(e) => match e.kind() {
         io::ErrorKind::AlreadyExists => {}
         _ => panic!("{}", e),
       },
     };
-    let _ = match fs::create_dir_all(Path::new(&_self.temp).join("encode")) {
+    let _ = match fs::create_dir_all(Path::new(&self.temp).join("encode")) {
       Ok(_) => {}
       Err(e) => match e.kind() {
         io::ErrorKind::AlreadyExists => {}
@@ -1880,24 +1711,24 @@ impl Project {
       },
     };
 
-    set_log(&_self.logging).unwrap();
+    set_log(&self.logging).unwrap();
 
-    let splits = _self.split_routine(py);
+    let splits = self.split_routine();
 
-    let chunk_queue = _self.load_or_gen_chunk_queue(splits, py);
+    let chunk_queue = self.load_or_gen_chunk_queue(splits);
 
-    let done_path = Path::new(&_self.temp).join("done.json");
+    let done_path = Path::new(&self.temp).join("done.json");
 
     let mut initial_frames: usize = 0;
 
-    if _self.resume && done_path.exists() {
+    if self.resume && done_path.exists() {
       let _ = log("Resuming...");
 
       let done: DoneJson = serde_json::from_str(&fs::read_to_string(&done_path).unwrap()).unwrap();
       initial_frames = done.done.iter().map(|(_, frames)| frames).sum();
       let _ = log(format!("Resmued with {} encoded clips done", done.done.len()).as_str());
     } else {
-      let total = _self.get_frames(py);
+      let total = self.get_frames();
       let mut done_file = fs::File::create(&done_path).unwrap();
       done_file
         .write_all(
@@ -1911,82 +1742,76 @@ impl Project {
         .unwrap();
     }
 
-    if !_self.resume {
-      extract_audio(&_self.input, &_self.temp, _self.audio_params.clone());
+    if !self.resume {
+      extract_audio(&self.input, &self.temp, self.audio_params.clone());
     }
 
-    if _self.workers == 0 {
-      _self.workers = determine_workers(&_self.encoder).unwrap() as usize;
+    if self.workers == 0 {
+      self.workers = determine_workers(Encoder::from_str(&self.encoder).unwrap()).unwrap() as usize;
     }
-    _self.workers = cmp::min(_self.workers, chunk_queue.len());
+    self.workers = cmp::min(self.workers, chunk_queue.len());
     println!(
       "Queue: {} Workers: {} Passes: {}\nParams: {}",
       chunk_queue.len(),
-      _self.workers,
-      _self.passes,
-      _self.video_params.join(" ")
+      self.workers,
+      self.passes,
+      self.video_params.join("\n")
     );
 
-    init_progress_bar((_self.frames - initial_frames) as u64).unwrap();
+    init_progress_bar((self.frames - initial_frames) as u64).unwrap();
 
-    Python::with_gil(|_| -> PyResult<()> {
-      // hack to avoid borrow checker errors
-      let concat = _self.concat.clone();
-      let temp = _self.temp.clone();
-      let input = _self.input.clone();
-      let output_file = _self.output_file.clone();
-      let encoder = _self.encoder.clone();
-      let vmaf = _self.vmaf;
-      let keep = _self.keep;
+    // hack to avoid borrow checker errors
+    let concat = self.concat.clone();
+    let temp = self.temp.clone();
+    let input = self.input.clone();
+    let output_file = self.output_file.clone();
+    let encoder = self.encoder.clone();
+    let vmaf = self.vmaf;
+    let keep = self.keep;
 
-      let queue = Queue {
-        chunk_queue,
-        project: &_self,
-        target_quality: if _self.target_quality.is_some() {
-          Some(TargetQuality::new(&_self))
-        } else {
-          None
-        },
-      };
+    let queue = Queue {
+      chunk_queue,
+      project: &self,
+      target_quality: if self.target_quality.is_some() {
+        Some(TargetQuality::new(&self))
+      } else {
+        None
+      },
+    };
 
-      queue.encoding_loop().unwrap();
+    queue.encoding_loop().unwrap();
 
-      let _ = log("Concatenating");
+    let _ = log("Concatenating");
 
-      // TODO refactor into Concatenate trait
-      match concat.as_str() {
-        "ivf" => {
-          av1an_core::concat::concat_ivf(&Path::new(&temp).join("encode"), Path::new(&output_file))
-            .unwrap();
-        }
-        "mkvmerge" => {
-          av1an_core::concat::concatenate_mkvmerge(temp.clone(), output_file.clone()).unwrap()
-        }
-        "ffmpeg" => {
-          av1an_core::ffmpeg::concatenate_ffmpeg(
-            temp.clone(),
-            output_file.clone(),
-            Encoder::from_str(&encoder).unwrap(),
-          );
-        }
-        _ => unreachable!(),
+    // TODO refactor into Concatenate trait
+    match concat.as_str() {
+      "ivf" => {
+        av1an_core::concat::concat_ivf(&Path::new(&temp).join("encode"), Path::new(&output_file))
+          .unwrap();
       }
-
-      if vmaf {
-        plot_vmaf(&input, &output_file)?;
+      "mkvmerge" => {
+        av1an_core::concat::concatenate_mkvmerge(temp.clone(), output_file.clone()).unwrap()
       }
-
-      if !keep {
-        fs::remove_dir_all(temp).unwrap();
+      "ffmpeg" => {
+        av1an_core::ffmpeg::concatenate_ffmpeg(
+          temp.clone(),
+          output_file.clone(),
+          Encoder::from_str(&encoder).unwrap(),
+        );
       }
+      _ => unreachable!(),
+    }
 
-      Ok(())
-    })
-    .unwrap();
+    if vmaf {
+      plot_vmaf(&input, &output_file).unwrap();
+    }
+
+    if !keep {
+      fs::remove_dir_all(temp).unwrap();
+    }
   }
 }
 
-#[pyfunction]
 fn run_vmaf_on_chunk(
   encoded: String,
   pipe_cmd: Vec<String>,
@@ -2011,68 +1836,4 @@ fn run_vmaf_on_chunk(
     threads,
   )
   .unwrap()
-}
-#[pymodule]
-fn av1an_pyo3(_py: Python, m: &PyModule) -> PyResult<()> {
-  m.add_function(wrap_pyfunction!(init_progress_bar, m)?)?;
-  m.add_function(wrap_pyfunction!(update_bar, m)?)?;
-  m.add_function(wrap_pyfunction!(finish_progress_bar, m)?)?;
-  m.add_function(wrap_pyfunction!(get_ffmpeg_info, m)?)?;
-  m.add_function(wrap_pyfunction!(determine_workers, m)?)?;
-  m.add_function(wrap_pyfunction!(create_vs_file, m)?)?;
-  m.add_function(wrap_pyfunction!(hash_path, m)?)?;
-  m.add_function(wrap_pyfunction!(adapt_probing_rate, m)?)?;
-  m.add_function(wrap_pyfunction!(frame_probe_vspipe, m)?)?;
-  m.add_function(wrap_pyfunction!(ffmpeg_get_frame_count, m)?)?;
-  m.add_function(wrap_pyfunction!(get_keyframes, m)?)?;
-  m.add_function(wrap_pyfunction!(concatenate_ivf, m)?)?;
-  m.add_function(wrap_pyfunction!(construct_target_quality_command, m)?)?;
-  m.add_function(wrap_pyfunction!(concatenate_ffmpeg, m)?)?;
-  m.add_function(wrap_pyfunction!(extract_audio, m)?)?;
-  m.add_function(wrap_pyfunction!(get_frame_types, m)?)?;
-  m.add_function(wrap_pyfunction!(extra_splits, m)?)?;
-  m.add_function(wrap_pyfunction!(segment, m)?)?;
-  m.add_function(wrap_pyfunction!(process_inputs, m)?)?;
-  m.add_function(wrap_pyfunction!(write_scenes_to_file, m)?)?;
-  m.add_function(wrap_pyfunction!(read_scenes_from_file, m)?)?;
-  m.add_function(wrap_pyfunction!(parse_args, m)?)?;
-  m.add_function(wrap_pyfunction!(default_args, m)?)?;
-  m.add_function(wrap_pyfunction!(vmaf_auto_threads, m)?)?;
-  m.add_function(wrap_pyfunction!(set_log, m)?)?;
-  m.add_function(wrap_pyfunction!(log, m)?)?;
-  m.add_function(wrap_pyfunction!(get_default_pass, m)?)?;
-  m.add_function(wrap_pyfunction!(get_default_cq_range, m)?)?;
-  m.add_function(wrap_pyfunction!(get_default_arguments, m)?)?;
-  m.add_function(wrap_pyfunction!(help_command, m)?)?;
-  m.add_function(wrap_pyfunction!(encoder_bin, m)?)?;
-  m.add_function(wrap_pyfunction!(output_extension, m)?)?;
-  m.add_function(wrap_pyfunction!(compose_ffmpeg_pipe, m)?)?;
-  m.add_function(wrap_pyfunction!(compose_1_1_pass, m)?)?;
-  m.add_function(wrap_pyfunction!(compose_1_2_pass, m)?)?;
-  m.add_function(wrap_pyfunction!(compose_2_2_pass, m)?)?;
-  m.add_function(wrap_pyfunction!(find_aom_keyframes, m)?)?;
-  m.add_function(wrap_pyfunction!(man_command, m)?)?;
-  m.add_function(wrap_pyfunction!(match_line, m)?)?;
-  m.add_function(wrap_pyfunction!(weighted_search, m)?)?;
-  m.add_function(wrap_pyfunction!(probe_cmd, m)?)?;
-  m.add_function(wrap_pyfunction!(get_percentile, m)?)?;
-  m.add_function(wrap_pyfunction!(read_weighted_vmaf, m)?)?;
-  m.add_function(wrap_pyfunction!(plot_vmaf_score_file, m)?)?;
-  m.add_function(wrap_pyfunction!(validate_vmaf, m)?)?;
-  m.add_function(wrap_pyfunction!(plot_vmaf, m)?)?;
-  m.add_function(wrap_pyfunction!(interpolate_target_q, m)?)?;
-  m.add_function(wrap_pyfunction!(interpolate_target_vmaf, m)?)?;
-  m.add_function(wrap_pyfunction!(log_probes, m)?)?;
-  m.add_function(wrap_pyfunction!(av_scenechange_detect, m)?)?;
-  m.add_function(wrap_pyfunction!(is_vapoursynth, m)?)?;
-  m.add_function(wrap_pyfunction!(save_chunk_queue, m)?)?;
-  m.add_function(wrap_pyfunction!(read_chunk_queue, m)?)?;
-  m.add_function(wrap_pyfunction!(frame_probe, m)?)?;
-  m.add_function(wrap_pyfunction!(run_vmaf_on_chunk, m)?)?;
-
-  m.add_class::<Project>()?;
-  m.add_class::<Chunk>()?;
-  m.add_class::<TargetQuality>()?;
-
-  Ok(())
 }
