--- av1an.py
@@ -1,6 +1,0 @@
-#!/usr/bin/env python3
-
-if __name__ == "__main__":
-    from cli.__main__ import main
-
-    main()
