--- av1an/__init__.py
@@ -1,4 +1,0 @@
-#!/usr/bin/env python3
-from av1an_pyo3 import *
-
-from av1an import *
