--- av1an/arg_parse.py
@@ -1,60 +1,0 @@
-#!/bin/env python
-import json
-import sys
-import shlex
-from pathlib import Path
-
-from av1an_pyo3 import (
-    default_args,
-    parse_args,
-    hash_path,
-    get_default_pass,
-    Project,
-    is_vapoursynth,
-)
-
-
-class Args:
-    def __init__(self):
-        self.defaults = json.loads(default_args())
-
-    def get_project(self):
-        self.parsed = json.loads(parse_args())
-        self.parsed["ffmpeg"] = self.parsed["ffmpeg"] if self.parsed["ffmpeg"] else ""
-        if self.parsed["temp"] is None:
-            self.parsed["temp"] = f".{hash_path(self.parsed['input'])}"
-        self.parsed["mkvmerge"] = False
-        self.parsed["output_ivf"] = False
-
-        if self.parsed["passes"] is None:
-            self.parsed["passes"] = get_default_pass(self.parsed["encoder"])
-
-        if self.parsed["video_params"] is None:
-            self.parsed["video_params"] = []
-        else:
-            self.parsed["video_params"] = shlex.split(self.parsed["video_params"])
-
-        if self.parsed["output_file"] is None:
-            self.parsed[
-                "output_file"
-            ] = f"{Path(self.parsed['input']).stem}_{self.parsed['encoder']}.mkv"
-
-        self.parsed["audio_params"] = shlex.split(self.parsed["audio_params"])
-        self.parsed["ffmpeg_pipe"] = []
-        if self.parsed["logging"] is None:
-            self.parsed["logging"] = (
-                (Path(self.parsed["temp"]) / "log.log").resolve().as_posix()
-            )
-        else:
-            self.parsed["logging"] = (
-                (Path(self.parsed["logging"]).with_suffix(".log")).resolve().as_posix()
-            )
-
-        self.parsed["frames"] = 0
-        self.parsed["is_vs"] = is_vapoursynth(self.parsed["input"])
-
-        if not self.parsed["input"]:
-            print("No input")
-            sys.exit()
-
-        return Project(self.parsed)
